backend/requirements.txt
anyio==3.7.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
Deprecated==1.2.14
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
jaconv==0.3.4
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
mecab-python3
orjson==3.9.1
pycparser==2.21
pydantic==1.10.9
pykakasi==2.2.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
srt==3.5.3
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.8.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
wrapt==1.15.0
